Rename interrupt_manager to interrupts
